feat: set context menu on parent element
