+++ src/components/ChatPreview/ChatPreview.tsx
@@ -1,6 +1,4 @@
 /* eslint-disable no-undef */
-import { useState, useEffect } from 'react';
-import ContextMenu from '../ContextMenu/ContextMenu';
 import Avatar from '../Avatar/Avatar';
 import type { UserData } from '../../types';
 import './ChatPreview.scss';
@@ -20,44 +18,24 @@ function ChatPreview({
 
   const selectChat = () => setActiveChatId(id);
 
-  const [showMenu, setShowMenu] = useState(false);
-
-  const handleContextMenu = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-    setShowMenu(true);
-  };
-
-  const closeContextMenu = () => setShowMenu(false);
-
-  useEffect(() => {
-    document.addEventListener('click', closeContextMenu);
-    return () => {
-      document.removeEventListener('click', closeContextMenu);
-    };
-  }, []);
-
   return (
-    <>
-      <button
-        type="button"
-        className={`chat-preview ${isActive ? 'active' : ''}`}
-        onClick={selectChat}
-        onContextMenu={handleContextMenu}
-      >
-        <div className="chat-preview-wrapper">
-          <Avatar image={avatar} />
-          <div className="chat-preview-text">
-            <div className="chat-preview__title">{name}</div>
-            <div className="chat-preview__last-message">{lastMessage}</div>
-          </div>
+    <button
+      type="button"
+      className={`chat-preview ${isActive ? 'active' : ''}`}
+      onClick={selectChat}
+    >
+      <div className="chat-preview-wrapper">
+        <Avatar image={avatar} />
+        <div className="chat-preview-text">
+          <div className="chat-preview__title">{name}</div>
+          <div className="chat-preview__last-message">{lastMessage}</div>
         </div>
-        <div className="chat-preview__info">
-          <div className="chat-preview__messenge-time">{timeOfLastMessage}</div>
-          {unreadMessages ? <div className="chat-preview__messenge-num">{unreadMessages}</div> : ''}
-        </div>
-      </button>
-      <ContextMenu isVisible={showMenu} handleMouseLeave={closeContextMenu} />
-    </>
+      </div>
+      <div className="chat-preview__info">
+        <div className="chat-preview__messenge-time">{timeOfLastMessage}</div>
+        {unreadMessages ? <div className="chat-preview__messenge-num">{unreadMessages}</div> : ''}
+      </div>
+    </button>
   );
 }
