+++ src/components/ContextMenu/ContextMenu.scss
@@ -4,6 +4,7 @@
     position: fixed;
     top: 30%;
     left: 10%;
+    z-index: 10;
 
     padding: 8px;
 
@@ -12,13 +13,13 @@
     box-shadow: 0px 0px 5px 2px rgb(0 0 0 / 22%);
     opacity: 0;
 
-    transition: linear .2s;
+    // transition: linear .2s;
 
-    visibility: hidden;
+    display: none;
 
     &.active {
         opacity: 1;
 
-        visibility: visible;
+        display: block;
     }
 }
