+++ src/components/ContextMenu/ContextMenu.tsx
@@ -5,16 +5,18 @@ import './ContextMenu.scss';
 
 interface ContextMenuProps {
   isVisible: boolean,
-  handleMouseLeave: () => void
+  handleMouseLeave: () => void,
+  position: { x: number, y: number };
 }
 
-function ContextMenu({ isVisible, handleMouseLeave }: ContextMenuProps) {
+function ContextMenu({ isVisible, handleMouseLeave, position }: ContextMenuProps) {
   const click = () => null;
 
   return (
     <nav
       className={`context-menu ${isVisible ? 'active' : ''}`}
       onMouseLeave={handleMouseLeave}
+      style={{ position: 'fixed', top: `${position.y}px`, left: `${position.x}px` }}
     >
       <PopupMenuItem label="Make as unread" icon={<SettingsIcon />} handleClick={click} />
       <PopupMenuItem label="Delete chat" icon={<SettingsIcon />} handleClick={click} />
